refactor: improved api to make more modular
