--- deboa-core/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "deboa-core"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-http = "1.3.1"
-strum_macros = "0.27.2"
