--- deboa-core/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod middlewares;
-pub mod request;
-pub mod response;
