+++ deboa-extras/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "deboa-extras"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+deboa = { path = "../deboa" }
+rmp-serde = { version = "1.3.0", optional = true }
+serde = { version = "1.0.217", features = ["derive"], optional = true }
+serde-xml-rust = { version = "0.6.0", optional = true }
+serde_json = { version = "1.0.138", optional = true }
+brotli = { version = "8.0.2", optional = true }
+flate2 = { version = "1.1.2", optional = true }
+bytes = "1.10.1"
+http = "1.3.1"
+mime = "0.3.17"
+tokio = "1.47.1"
+httpmock = "0.7.0"
+
+[features]
+default = ["json", "compression"]
+# serializarion
+json = ["dep:serde_json", "dep:serde"]
+xml = ["dep:serde-xml-rust", "dep:serde"]
+msgpack = ["dep:rmp-serde", "dep:serde"]
+# compression
+compression = ["brotli"]
+brotli = ["dep:brotli"]
+deflate = ["dep:flate2"]
+gzip = ["dep:flate2"]
+
+[dev-dependencies]
+httpmock = "0.7.0"
