+++ deboa-extras/src/compression/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "brotli")]
+pub mod brotli;
+#[cfg(feature = "deflate")]
+pub mod deflate;
+#[cfg(feature = "gzip")]
+pub mod gzip;
