+++ deboa-extras/src/serialization/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "json")]
+pub mod json;
+
+#[cfg(feature = "xml")]
+pub mod xml;
+
+#[cfg(feature = "msgpack")]
+pub mod msgpack;
