+++ deboa-extras/src/tests/mod.rs
@@ -1,0 +1,8 @@
+mod types;
+
+#[cfg(feature = "json")]
+mod json;
+#[cfg(feature = "msgpack")]
+mod msgpack;
+#[cfg(feature = "xml")]
+mod xml;
