+++ deboa/Cargo.toml
@@ -10,19 +10,10 @@ license = "GPL-3.0-or-later"
 keywords = ["http", "networking", "client", "fetch", "rest-client"]
 
 [features]
-default = ["http1", "middlewares", "json", "tokio-rt"]
+default = ["http1", "middlewares", "tokio-rt"]
 middlewares = []
 http1 = []
 http2 = []
-# serializarion
-json = ["dep:serde_json", "dep:serde"]
-xml = ["dep:serde-xml-rust", "dep:serde"]
-msgpack = ["dep:rmp-serde", "dep:serde"]
-# compression
-compression = ["brotli"]
-brotli = ["dep:brotli"]
-deflate = ["dep:flate2"]
-gzip = ["dep:flate2"]
 # runtime
 tokio-rt = ["dep:tokio"]
 smol-rt = [
@@ -44,8 +35,6 @@ compio-rt = [
 anyhow = { version = "1.0.99", optional = true }
 async-native-tls = { version = "0.5.0", optional = true }
 base64 = "0.22.1"
-flate2 = { version = "1.1.2", optional = true }
-brotli = { version = "8.0.2", optional = true }
 bytes = { version = "1.2", optional = false }
 compio = { version = "0.15.0", optional = true, features = ["macros"] }
 cyper-core = { version = "0.4.0", optional = true, features = ["client"] }
@@ -62,10 +51,7 @@ hyper-util = { version = "0.1.10", features = [
 ] }
 macro_rules_attribute = { version = "0.2.2", optional = true }
 mime = { version = "0.3.17" }
-rmp-serde = { version = "1.3.0", optional = true }
 serde = { version = "1.0.217", features = ["derive"], optional = true }
-serde-xml-rust = { version = "0.6.0", optional = true }
-serde_json = { version = "1.0.138", optional = true }
 smol = { version = "2.0.2", optional = true }
 smol-hyper = { version = "0.1.0", optional = true }
 smol-macros = { version = "0.1.1", optional = true }
