+++ examples/simple/Cargo.toml
@@ -5,7 +5,8 @@ edition = "2021"
 publish = false
 
 [dependencies]
-deboa = { path = "../../deboa", features = ["http1", "json", "tokio-rt"] }
+deboa = { path = "../../deboa", features = ["http1", "tokio-rt"] }
+deboa-extras = { path = "../../deboa-extras", features = ["json"] }
 tokio = { version = "1.47.1", features = ["full"] }
 serde = "1.0.219"
 serde_json = "1.0.138"
